moved the key value for the book display component from the component itself to the component representation in the BookListView.
added functionality to allow for switching between displaying for
search results in books and authors by changing the search type state.
The useEffect refreshes the page everytime the search type changes.

Added functionality on user profile to update the user "about" by
creating a corelated firebase database that stores the "about" under
uids.
the profile component checks on mount for the existence of the document
that stores the "about" and if not displays an empty section.
The "update information" button displays a text area that allows to
 add/update the "about" section and the "save data" button adds the
 information to the firebase database if there is a change from
  the previous "about" value.